+++ src/Serialization/TablePayloadTypes.php
@@ -60,6 +60,27 @@ namespace Mheads\Yii\Table\Serialization;
  *     columnKey: string|null,
  *     ...
  * }
+ * @psalm-type TableConfigPayload=array{
+ *     config: array{
+ *         tableId: string,
+ *         filterParam: string,
+ *         sortParam: string|null,
+ *         pageParam: string|null,
+ *         pageSizeParam: string|null,
+ *         pageSizeConstraint: bool|int|array<int, int>|null,
+ *         prevPageParam: string|null,
+ *         columnIdKey: string|null,
+ *         exportParam: string|null,
+ *         exportCodes: array<int, string>|null
+ *     },
+ *     columns: array<int, ColumnPayload>,
+ *     filters: array<int, array<array-key, mixed>>,
+ *     sorts: array<int, SortOptionPayload>
+ * }
+ * @psalm-type TableRowsPayload=array{
+ *     pagination: OffsetPagination|KeysetPagination|GenericPagination|null,
+ *     rows: array<int, array<string, mixed>>
+ * }
  * @psalm-type TablePayload=array{
  *     config: array{
  *         tableId: string,